Refactoring
Refactoring the structor of the code
